refactor: inventory マクロによるプラグイン自動登録で registry.rs を削除 (#184)
- `inventory = "0.3"` を Cargo.toml に追加
- `PluginRegistration` 型と `inventory::collect!` を `plugins/mod.rs` に定義
- `build_registry()` / `find_plugin()` を `mod.rs` に移動し `registry.rs` を削除
- `DmmPlugin` / `HobbySearchPlugin` に `inventory::submit!` を追加
- サブモジュールを `pub mod` に変更して LTO によるデッドコード除外を防止
- `registry.rs` のテスト 11 件を `plugins::tests` に移行
- README.md・docs/architecture/plugin-system.md を Phase 3 対応内容に更新
